requirements.txt
selenium
psutil
pytest
pytest-cov
pytest-xdist
allure-pytest
selenium-page-factory